Fix volume being an int

+++ nt3window.cpp
@@ -2,7 +2,7 @@
 
 double framerate;
 
-double volume;
+int volume;
 double hue;
 bool fullscreen;
